+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.2] - 2026-03-15
+
 ### Added
 
 - Add `validatePortAvailable` function to `src/main/services/code-server/path-mapper.ts` to double-check port availability after initial discovery.